requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
joblib
scikit-learn
comet-ml
dvc